+++ src/topbar/settingDialog.tsx
@@ -9,21 +9,8 @@ import {
 } from '../token';
 import { Widget } from '@lumino/widgets';
 import type { AppWidget } from '../specta_widget';
-import type { AppModel } from '../specta_model';
+import { StaticRenderingSection } from './staticRenderingSection';
 
-type ISnapshotState = {
-  status: 'out-of-sync' | 'in-sync' | 'not-exist';
-  timestamp?: number;
-  staticRender: boolean;
-};
-
-function readSnapshotState(model?: AppModel): ISnapshotState {
-  return {
-    status: model?.snapshotStatus() ?? 'not-exist',
-    timestamp: model?.getSnapshot()?.timestamp,
-    staticRender: Boolean(model?.staticRender)
-  };
-}
 export const SettingContent = (props: {
   config?: ITopbarConfig;
   themeManager?: IThemeManager;
@@ -151,60 +138,6 @@ export const SettingContent = (props: {
     [uiSwitcher, currentPath]
   );
 
-  const model = props.spectaWidget?.model;
-  const [snapshotState, setSnapshotState] = useState<ISnapshotState>(() =>
-    readSnapshotState(model)
-  );
-  const [creatingSnapshot, setCreatingSnapshot] = useState(false);
-
-  // Names kept so the JSX below is untouched.
-  const {
-    status: snapshotStatus,
-    timestamp: currentTimestamp,
-    staticRender: isStaticRendering
-  } = snapshotState;
-
-  useEffect(() => {
-    if (!model) {
-      return;
-    }
-    const handler = () => setSnapshotState(readSnapshotState(model));
-    model.snapshotChanged.connect(handler);
-    return () => {
-      model.snapshotChanged.disconnect(handler);
-    };
-  }, [model]);
-
-  const creatingRef = useRef(false);
-
-  const deleteSnapshot = useCallback(async () => {
-    if (snapshotStatus === 'not-exist') {
-      return;
-    }
-    await model?.deleteSnapshot();
-  }, [model, snapshotStatus]);
-
-  const createSnapshot = useCallback(async () => {
-    if (isStaticRendering || creatingRef.current) {
-      return;
-    }
-    creatingRef.current = true;
-    setCreatingSnapshot(true);
-    try {
-      await props.spectaWidget?.saveSnapshot();
-    } finally {
-      creatingRef.current = false;
-      setCreatingSnapshot(false);
-    }
-  }, [props.spectaWidget, isStaticRendering]);
-
-  const activateKernel = useCallback(async () => {
-    if (!isStaticRendering) {
-      return;
-    }
-    await props.spectaWidget?.turnOffStaticRender();
-  }, [props.spectaWidget, isStaticRendering]);
-
   return (
     <div style={{ padding: '0 10px' }}>
       <p style={{ marginTop: 0, marginBottom: '5px', fontSize: '1rem' }}>
@@ -298,100 +231,10 @@ export const SettingContent = (props: {
         </div>
       )}
       {props.enableStaticRenderingConfig && (
-        <div>
-          <label htmlFor="">
-            <b>Static rendering: {isStaticRendering ? 'On' : 'Off'}</b>
-          </label>
-          <div
-            style={{
-              marginBottom: '12px',
-              display: 'flex',
-              gap: '2px',
-              flexDirection: 'column'
-            }}
-          >
-            {snapshotStatus !== 'not-exist' && (
-              <div
-                style={{ flexGrow: 1, display: 'flex', alignItems: 'center' }}
-              >
-                Last snapshot:{' '}
-                {currentTimestamp
-                  ? new Date(currentTimestamp).toLocaleString()
-                  : 'Unavailable'}
-              </div>
-            )}
-            <div style={{ flexGrow: 1, display: 'flex', alignItems: 'center' }}>
-              {snapshotStatus === 'not-exist'
-                ? 'No snapshot found'
-                : snapshotStatus === 'out-of-sync'
-                  ? 'Snapshot is out of sync with the notebook'
-                  : ''}
-            </div>
-
-            <div
-              style={{
-                gap: '8px',
-                flexDirection: 'row',
-                marginBottom: '4px',
-                display: props.isSpectaApp ? 'none' : 'flex'
-              }}
-            >
-              <button
-                className="jp-mod-styled jp-mod-warn"
-                onClick={deleteSnapshot}
-                disabled={snapshotStatus === 'not-exist'}
-                style={{
-                  cursor:
-                    snapshotStatus === 'not-exist' ? 'not-allowed' : 'pointer',
-                  flexGrow: 1,
-                  opacity: snapshotStatus === 'not-exist' ? 0.5 : 1
-                }}
-              >
-                Delete snapshot
-              </button>
-              <button
-                className="jp-mod-styled jp-mod-accept"
-                onClick={createSnapshot}
-                style={{
-                  cursor:
-                    isStaticRendering || creatingSnapshot
-                      ? 'not-allowed'
-                      : 'pointer',
-                  flexGrow: 1,
-                  opacity: isStaticRendering || creatingSnapshot ? 0.5 : 1
-                }}
-                disabled={isStaticRendering || creatingSnapshot}
-                title={
-                  isStaticRendering
-                    ? 'Cannot create snapshot in static rendering mode'
-                    : ''
-                }
-              >
-                {creatingSnapshot ? 'Creating...' : 'Create snapshot'}
-              </button>
-            </div>
-            <div
-              style={{
-                display: props.spectaWidget ? 'flex' : 'none',
-                justifyContent: 'center'
-              }}
-            >
-              <button
-                className="jp-mod-styled jp-mod-accept"
-                onClick={activateKernel}
-                disabled={!isStaticRendering}
-                style={{
-                  cursor: !isStaticRendering ? 'not-allowed' : 'pointer',
-                  flexGrow: 1,
-                  opacity: !isStaticRendering ? 0.5 : 1
-                }}
-                title="Render notebook using a live kernel"
-              >
-                Render with kernel
-              </button>
-            </div>
-          </div>
-        </div>
+        <StaticRenderingSection
+          spectaWidget={props.spectaWidget}
+          isSpectaApp={props.isSpectaApp}
+        />
       )}
       {settingsWidgets && settingsWidgets.length > 0 && (
         <div className="specta-settings-custom-section">
